Add feature toggle
