+++ src/pages/api/simple-feature-toggle/index.ts
@@ -4,6 +4,7 @@ import { authoriseServiceAPIRequest } from '@/utils/serviceApiClient'
 export interface SimpleFeatureToggleResponse {
   googleTagManagerEnabled: boolean
   enableFollowOnIsEmergencyField: boolean
+  enableNewAwaabsFields: boolean
 }
 
 export default authoriseServiceAPIRequest(async (req, res) => {
@@ -12,6 +13,8 @@ export default authoriseServiceAPIRequest(async (req, res) => {
 
     enableFollowOnIsEmergencyField:
       process.env.FOLLOW_ON_IS_EMERGENCY_FIELD_ENABLED === 'true',
+
+    enableNewAwaabsFields: process.env.ENABLE_NEW_AWAABS_FIELDS === 'true',
   }
 
   res.status(HttpStatus.StatusCodes.OK).json(data)
